Add files via upload
Group project for my Organisation Behavior Module. The project was based on providing analysis to a hypothetical company in the evaluation of employee performance, job satisfaction and motivation, we worked to provide a solution at the end of this project. I achieved an A grade for my contribution to the project.