Add files via upload
Java Swing thread timer. This desktop app involved concurrency control using threads. A user could start, stop and pause a timer using buttons. Each time the user clicked a button the action was recorded and displayed in a separate window. Users could save the information to a file and open it later on.